pyproject.toml
[project]
name = "reeeductio"
version = "0.1.0"
description = "End-to-end encrypted PubSub messaging system with capability-based authorization"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.0.0",
    "pyjwt>=2.8.0",
    "cryptography>=42.0.0",
    "python-multipart>=0.0.6",
    "boto3>=1.34.0",
    "pyyaml>=6.0.0",
    "google-cloud-firestore>=2.14.0",  # For Firestore backend support
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-xdist>=3.5.0",  # For parallel test execution
    "pytest-asyncio>=0.23.0",  # For async test support
    "httpx>=0.25.0",  # For WebSocket testing with TestClient
    "testcontainers[google]>=3.7.0",  # For Firestore emulator in tests
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = [
    "main.py",
    "database.py",
    "crypto.py",
    "authorization.py",
    "example_client.py",
]

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
